Create testUrlArray
Three versions of using an array to test the validity of a URL. intended to replace (or supplement) the role of the line "check('giturl').contains('sts')" in patController.js. I prefer the third version, which tests using a supplemental MongoDB file that can be updated independently of the code.